+++ cli/src/commands/git/remote/set_url.rs
@@ -59,7 +59,7 @@ pub async fn cmd_git_remote_set_url(
     let workspace_command = command.workspace_helper(ui).await?;
 
     let process_url = |url: Option<&String>| {
-        url.map(|url| absolute_git_url(command.cwd(), url))
+        url.map(|url| absolute_git_url(command.cwd(), url).map(|(url, _)| url))
             .transpose()
     };